pyproject.toml
[project]
name = "audit-clearance"
version = "0.1.0"
description = "Auditing agents with hidden objectives in sensitive information scenarios"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.102.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "typing-extensions>=4.15.0",
]

[tool.hatch.build.targets.wheel]
packages = ["arena"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
